introduce IAuthManager interface for configuring authentication; provide default implementation using Apache Shiro (Fixes issue 6)

+++ roblog/src/controller/AdminCx.gs
@@ -8,23 +8,27 @@ uses ronin.RoninController
 class AdminCx extends RoninController {
 
   static function newPost() {
-    view.Layout.render(Writer, Session["User"] as String, "New post", \ -> view.EditPost.render(Writer, new Post()))
+    view.Layout.render(Writer, AuthManager.CurrentUser, "New post", \ -> view.EditPost.render(Writer, new Post()))
   }
 
   static function editPost(post : Post) {
-    view.Layout.render(Writer, Session["User"] as String, "Edit post", \ -> view.EditPost.render(Writer, post))
+    view.Layout.render(Writer, AuthManager.CurrentUser, "Edit post", \ -> view.EditPost.render(Writer, post))
   }
 
   static function deletePost(post : Post) {
-    post.delete()
+    if(AuthManager.CurrentUser == "admin") {
+      post.delete()
+    }
     redirect(PostCx#recent(0))
   }
 
   static function savePost(post : Post) {
-    if(post._New) {
-      post.Posted = new java.sql.Timestamp(java.lang.System.currentTimeMillis())
+    if(AuthManager.CurrentUser == "admin") {
+      if(post._New) {
+        post.Posted = new java.sql.Timestamp(java.lang.System.currentTimeMillis())
+      }
+      post.update()
     }
-    post.update()
     redirect(PostCx#viewPost(post))
   }
 
@@ -36,54 +40,38 @@ class AdminCx extends RoninController {
     } else {
       blogInfo = blogInfos[0]
     }
-    view.Layout.render(Writer, Session["User"] as String, "Setup", \ -> view.Setup.render(Writer, blogInfo))
+    view.Layout.render(Writer, AuthManager.CurrentUser, "Setup", \ -> view.Setup.render(Writer, blogInfo))
   }
 
   static function editInfo(blogInfo : BlogInfo) {
-    blogInfo.update()
+    if(AuthManager.CurrentUser == "admin") {
+      blogInfo.update()
+    }
     redirect(#setup())
   }
 
   static function login() {
-    view.Layout.render(Writer, Session["User"] as String, "Login", \ -> view.Login.render(Writer))
+    view.Layout.render(Writer, AuthManager.CurrentUser, "Login", \ -> view.Login.render(Writer))
   }
 
   static function doLogin(name : String, pass : String) {
-    var user = User.find(new User() {:Name = name})[0]
-    if(user != null) {
-        if(User.Hash == new String(hashPass(pass.getBytes(), User.Salt.getBytes()))) {
-            Session["User"] = User.Name
-            redirect(PostCx#recent(0))
-            return
-        }
+    if(AuthManager.login(name, pass)) {
+      redirect(PostCx#recent(0))
+    } else {
+      redirect(#login())
     }
-    redirect(#login())
   }
 
   static function logout() {
-    Session["User"] = null
+    AuthManager.logout()
     redirect(#login())
   }
 
-  /*
-      ATTENTION COPY-PASTERS:
-      Please do not use the code below for your project.  Use a real authentication system.
-      This is a toy example and probably very easily hacked.
-  */
-
-  static function /*DONOTCOPYPASTE*/createUser(name : String, pass : String) {
-    var salt = new byte[32]
-    java.security.SecureRandom.getInstance("SHA1PRNG").nextBytes(salt)
-    var passBytes = pass.getBytes()
-    var user = new User() {:Name = name, :Hash = new String(hashPass(passBytes, salt)), :Salt = new String(salt)}
+  static function createUser(name : String, pass : String) {
+    var hashAndSalt = AuthManager.getPasswordHashAndSalt(pass)
+    var user = new User() {:Name = name, :Hash = hashAndSalt.First, :Salt = hashAndSalt.Second}
     user.update()
     redirect(#login())
   }
 
-  static function /*DONOTCOPYPASTE*/hashPass(pass : byte[], salt : byte[]) : byte[] {
-    var saltedPass = new byte[salt.length + pass.length]
-    java.lang.System.arrayCopy(salt, 0, saltedPass, 0, salt.length)
-    java.lang.System.arrayCopy(pass, 0, saltedPass, salt.length, pass.length)
-    return java.security.MessageDigest.getInstance("MD5").digest(saltedPass)
-  }
 }
